packages/gooddata-sdk/pyproject.toml
# (C) 2025 GoodData Corporation
[project]
name = "gooddata-sdk"
version = "1.53.0"
description = "GoodData Cloud Python SDK"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "GoodData", email = "support@gooddata.com"}
]
keywords = [
    "gooddata",
    "sdk",
    "api",
    "analytics",
    "headless",
    "business",
    "intelligence",
    "headless-bi",
    "cloud",
    "native",
    "semantic",
    "layer",
    "sql",
    "metrics",
]
requires-python = ">=3.9.0"
dependencies = [
    "gooddata-api-client~=1.53.0",
    "python-dateutil>=2.5.3",
    "pyyaml>=6.0",
    "attrs>=21.4.0,<=24.2.0",
    "cattrs>=22.1.0,<=24.1.1",
    "brotli==1.1.0",
    "requests~=2.32.0",
    "python-dotenv>=1.0.0,<2.0.0",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Database",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development",
    "Typing :: Typed",
]

[project.urls]
Documentation = "https://www.gooddata.com/docs/python-sdk/1.53.0"
Source = "https://github.com/gooddata/gooddata-python-sdk"

[project.scripts]
gdc = "gooddata_sdk.cli.main:main"

[dependency-groups]
test = [
    "pytest~=8.3.4",
    "pytest-cov~=6.0.0",
    "pytest-snapshot==0.9.0",
    "pytest-order~=1.3.0",
    "vcrpy~=7.0.0",
    "urllib3==1.26.9",
    "python-dotenv~=1.0.0",
    "deepdiff~=8.5.0"
]
type = [
    "mypy~=1.11.2",
    "pydantic~=2.9.2",
    "types-python-dateutil >= 2.5.3",
    "types-pyyaml~=6.0.0",
    "attrs>=21.4.0,<=24.2.0",
    "cattrs>=22.1.0,<=24.1.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/gooddata_sdk"]
include = [
    "gooddata_sdk/cli/package.json",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
